Struct wasmtime_wmemcheck::Wmemcheck

pub struct Wmemcheck {
    pub stack_pointer: usize,
    pub flag: bool,
    /* private fields */
}

Memory checker for wasm guest.

Fields
stack_pointer: usizeflag: bool
Implementations

impl Wmemcheck

pub fn new(mem_size: usize) -> Wmemcheck

Initializes memory checker instance.

pub fn malloc(&mut self, addr: usize, len: usize) -> Result<(), AccessError>

Updates memory checker memory state metadata when malloc is called.

pub fn read(&mut self, addr: usize, len: usize) -> Result<(), AccessError>

Updates memory checker memory state metadata when a load occurs.

pub fn write(&mut self, addr: usize, len: usize) -> Result<(), AccessError>

Updates memory checker memory state metadata when a store occurs.

pub fn free(&mut self, addr: usize) -> Result<(), AccessError>

Updates memory checker memory state metadata when free is called.

pub fn update_stack_pointer(&mut self, new_sp: usize) -> Result<(), AccessError>

Updates memory checker metadata when stack pointer is updated.

pub fn memcheck_on(&mut self)

Turns memory checking on.

pub fn memcheck_off(&mut self)

Turns memory checking off.

pub fn set_stack_size(&mut self, stack_size: usize)

Initializes stack and stack pointer in memory checker metadata.

pub fn update_mem_size(&mut self, num_bytes: usize)

Updates memory checker metadata size when memory.grow is called.